Add Roles/UserHomeGroups tables, migrate IDs to bigint, seed SuperAdmin from env
- New Role entity (seeded: SuperAdmin=1, Admin=2)
- New UserHomeGroup many-to-many (User <-> HomeGroup)
- All entity IDs changed int -> long (bigint)
- User.Role string field replaced with RoleId FK
- SuperAdmin user (id=0) seeded from SUPERADMIN_EMAIL/PASSWORD env vars on startup
- Removed public /register endpoint
- docker-compose: added SUPERADMIN_EMAIL/PASSWORD env vars
- Fresh migration replacing InitialCreate

+++ HomeGroup.API/Controllers/AttendanceController.cs
@@ -14,7 +14,7 @@ public class AttendanceController(AppDbContext db) : ControllerBase
 {
     [HttpGet]
     public async Task<ActionResult<List<AttendanceResponse>>> GetByGroup(
-        [FromQuery] int groupId,
+        [FromQuery] long groupId,
         [FromQuery] DateOnly? from,
         [FromQuery] DateOnly? to)
     {
@@ -34,7 +34,7 @@ public class AttendanceController(AppDbContext db) : ControllerBase
     }
 
     [HttpGet("summary")]
-    public async Task<ActionResult<List<AttendanceSummary>>> GetSummary([FromQuery] int groupId)
+    public async Task<ActionResult<List<AttendanceSummary>>> GetSummary([FromQuery] long groupId)
     {
         var summary = await db.Attendances
             .Where(a => a.HomeGroupId == groupId)
